+++ common/persistence/sql/sqlplugin/mysql/db.go
@@ -29,8 +29,8 @@ import (
 	"github.com/go-sql-driver/mysql"
 	"github.com/jmoiron/sqlx"
 
+	"github.com/uber/cadence/common/persistence/sql/sqldriver"
 	"github.com/uber/cadence/common/persistence/sql/sqlplugin"
-	"github.com/uber/cadence/common/persistence/sql/sqlplugin/sqldriver"
 )
 
 type (
