+++ common/persistence/sql/sqlplugin/postgres/db.go
@@ -28,8 +28,8 @@ import (
 	"github.com/jmoiron/sqlx"
 	"github.com/lib/pq"
 
+	"github.com/uber/cadence/common/persistence/sql/sqldriver"
 	"github.com/uber/cadence/common/persistence/sql/sqlplugin"
-	"github.com/uber/cadence/common/persistence/sql/sqlplugin/sqldriver"
 )
 
 type (
